ref: Move macros to their own directory
This puts the macros in a subdirectory. I have included macros from 3
themes in total, which is a little bit messy, however by putting them in
a subdirectory I can move to consolidating them in a single file.

This includes the updates to all templates to reference the appropriate
directory.